analytics/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.30.0
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
pydantic==2.10.4
pydantic-settings==2.6.1
httpx==0.25.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
gunicorn==21.2.0 